refactor(bot-config-manifest): bind the credential service beside its config
The R9 file-size guard failed the previous commit: 17 added lines took
``bot_management_module.py`` from 999 to 1015, past the 1000-line cap. The
guard's own message says what to do — "each file should hold one cohesive
concern" — and for this provider that reading is not a workaround, it is
the answer.

``SourceCredentialService`` lives in ``core/bot_config_manifest/credentials/``
and now reads ``BotConfigManifestConfig``. Both of those belong to
``ManifestFetchModule``, whose docstring already explains that the manifest
config cluster lives there rather than in ``config_module`` for exactly this
reason: that file sits at the cap too. The provider was in the bot-management
module because it landed there with W4, and the distance is what let the
allowlist go unwired for a wave — the config was one module away, so passing
it was never the obvious thing to do.

So the provider moves whole, taking the three imports only it used. The
repository binding stays with its repository siblings, which is where a
reader looks for it. ``bot_management_module.py`` lands at 968 lines and
``manifest_fetch_module.py`` at 682, both under the cap without an allowlist
entry — the guard is meant to shrink that list, not grow it for a five-line
wiring fix.

The regression test moves with the provider, into a wiring file named for the
module that now owns it. It still asserts through the real injector, so it is
indifferent to which module binds: removing the ``endpoint_allow_hosts``
argument fails it in its new home exactly as it did in the old one.
