+++ src/backend/src/agentclaw/community/di/modules/bot_management_module.py
@@ -240,13 +240,6 @@ from agentclaw.community.core.repository.implementations.bot.manifest_content im
 from agentclaw.community.core.repository.implementations.bot.source_credential import (
     SourceCredentialRepository,
 )
-from agentclaw.community.api.source_credential_service import (
-    SourceCredentialServiceProtocol,
-)
-from agentclaw.community.core.bot_management.token_vault import TokenVault
-from agentclaw.community.core.bot_config_manifest.credentials.service import (
-    SourceCredentialService,
-)
 
 from agentclaw.community.core.repository.implementations.bot.config_manifest_apply import (
     BotConfigManifestApplyLockRepository,
@@ -973,43 +966,3 @@ class BotManagementModule(Module):
         self, svc: DataInitService
     ) -> DataInitServiceProtocol:
         return svc
-
-    @singleton
-    @provider
-    @inject
-    def _source_credential_service_protocol(
-        self,
-        repository: SourceCredentialRepositoryProtocol,
-        vault: TokenVault,
-        manifest_config: cfg.BotConfigManifestConfig,
-    ) -> SourceCredentialServiceProtocol:
-        """W3 (#1471): the profile decides the fail-closed posture.
-
-        Production columns (corp, community) refuse credential writes when
-        the vault has no master key — TokenVault's plaintext passthrough
-        is right for local, catastrophic for tenant tokens at rest. The
-        local/test columns keep the permissive default; corp_test runs the
-        Mist-backed vault anyway and benefits from the same guard.
-
-        The transport allowlist is the SAME value ``GuardedFetcher`` takes,
-        from the same ``user_config.bot_config_manifest`` block. It has to
-        reach here too: the endpoint guard refuses a host that resolves to a
-        private address, which is precisely what an internal object store
-        endpoint does — and without this the deployment has no way to say so.
-        The escape hatch existed on the service and nothing passed through it,
-        so a legitimately internal endpoint was unregistrable by any
-        configuration, and the one place the deployment declares such a host
-        governed the fetch road only.
-        """
-        from agentclaw.community.di.profile import DeployProfile
-
-        fail_closed = DeployProfile.detect() in (
-            DeployProfile.CORP,
-            DeployProfile.COMMUNITY,
-        )
-        return SourceCredentialService(
-            repository,
-            vault,
-            fail_closed=fail_closed,
-            endpoint_allow_hosts=manifest_config.fetch_transport_allowlist,
-        )
